Moved lib tasks to lib folder
