+++ Gruntfile.js
@@ -10,7 +10,7 @@ module.exports = function(grunt) {
         jshint: {
             all: [
                 'Gruntfile.js',
-                'tasks/*.js',
+                'tasks/**/*.js',
                 '<%= nodeunit.tests %>'
             ],
             options: {
@@ -101,6 +101,6 @@ module.exports = function(grunt) {
     grunt.loadNpmTasks('grunt-contrib-watch');
 
     grunt.registerTask('default', 'bust');
-    grunt.registerTask('bust', ['clean', 'copy', 'cacheBust']);
-    grunt.registerTask('test', ['clean', 'copy', 'cacheBust', 'nodeunit']);
+    grunt.registerTask('test', ['bust', 'nodeunit']);
+    grunt.registerTask('bust', ['clean', 'jshint', 'copy', 'cacheBust']);
 };
