+++ tasks/lib/utils.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var grunt = require('grunt');
 var crypto = require('crypto');
